+++ gaze_controlled_cursor_demo/__main__.py
@@ -1,3 +1,3 @@
-import app
+import app3
 
-app.run()
+app3.run()
