+++ gaze_controlled_cursor_demo/gaze_provider/__init__.py
@@ -1,5 +1,42 @@
 from pupil_labs.realtime_api.simple import discover_one_device
 from pupil_labs.real_time_screen_gaze.gaze_mapper import GazeMapper
+from pupil_labs.real_time_screen_gaze import marker_generator
+
+from .marker_widget import MarkerWidget
+from PySide6.QtCore import QPoint
+
+# TODO make detected markers a property
+
+
+class DummyGazeProvider:
+    def __init__(self):
+        # self.device = None
+        # self.gazeMapper = None
+        # self.surface_definition = None
+        # self.surface = None
+        # self.surface_to_global_transform = surface_to_global_transform
+        self.marker_widget = MarkerWidget()
+
+    @property
+    def connected(self):
+        return True
+
+    def connect(self):
+        return True
+
+    def receive(self):
+        """The device must be connected and a surface must be defined before calling this method."""
+        import time
+
+        gaze = QPoint(1000, 500)
+        return gaze, time.time()
+
+    @classmethod
+    def generate_marker(cls, marker_id):
+        return marker_generator.generate_marker(marker_id, flip_x=True, flip_y=True)
+
+    def close(self):
+        pass
 
 
 class GazeProvider:
@@ -8,7 +45,8 @@ class GazeProvider:
         self.gazeMapper = None
         self.surface_definition = None
         self.surface = None
-        surface_to_screen_transform = None
+
+        self.marker_widget = MarkerWidget()
 
     @property
     def connected(self):
@@ -28,12 +66,9 @@ class GazeProvider:
 
         calibration = self.device.get_calibration()
         self.gazeMapper = GazeMapper(calibration)
-
-        if self.surface_definition is not None:
-            self.updateSurface(
-                self.surface_definition["marker_verts"],
-                self.surface_definition["surface_size"],
-            )
+        self.surface = self.gazeMapper.add_surface(
+            self.marker_widget.getMarkerVerts(), self.marker_widget.getSurfaceSize()
+        )
 
         return True
 
@@ -52,15 +87,27 @@ class GazeProvider:
         assert self.device is not None
         assert self.surface is not None
 
+        frame, raw_gaze = self._receive_data_from_device()
+        mapped_gaze, detected_markers = self._map_gaze(frame, raw_gaze)
+        if mapped_gaze is not None:
+            mapped_gaze = self.marker_widget.surface_to_global_transform(*mapped_gaze)
+
+        return mapped_gaze, raw_gaze.timestamp_unix_seconds
+
+    def _receive_data_from_device(self):
         device_response = self.device.receive_matched_scene_video_frame_and_gaze(
             timeout_seconds=1 / 15
         )
 
         if device_response is None:
-            print("No frame and gaze")
-            return None
+            raise ValueError("No frame and gaze received")
 
         frame, gaze = device_response
+        return frame, gaze
+
+    def _map_gaze(self, frame, gaze):
+        assert self.surface is not None
+
         result = self.gazeMapper.process_frame(frame, gaze)
 
         detected_markers = [int(marker.uid.split(":")[-1]) for marker in result.markers]
@@ -71,3 +118,11 @@ class GazeProvider:
                 gaze = surface_gaze.x, surface_gaze.y
 
         return gaze, detected_markers
+
+    @classmethod
+    def generate_marker(cls, marker_id):
+        return marker_generator.generate_marker(marker_id, flip_x=True, flip_y=True)
+
+    def close(self):
+        if self.device is not None:
+            self.device.close()
